requirements.txt
annotated-types==0.5.0
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
django-cors-headers==3.14.0
django-user-agents==0.4.0
djangorestframework==3.14.0
idna==3.4
psycopg2==2.9.7
psycopg2-binary==2.9.7
pydantic==2.2.1
pydantic_core==2.6.1
PyJWT==2.8.0
python-decouple==3.8
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
typing_extensions==4.7.1
tzdata==2022.7
ua-parser==0.18.0
urllib3==1.26.15
user-agents==2.2.0
whitenoise==6.4.0
